+++ helpers/documentHelper.js
@@ -47,6 +47,16 @@ documentHelper.isEditable = function (addon, extension) {
     return false;
 }
 
+documentHelper.isFillable = function (addon, extension) {
+    const formats = addon.config.docServer().formats || [];
+
+    for (let i = 0; i < formats.length; i++) {
+        if (formats[i].name === extension) return formats[i].actions.includes('fill');
+    }
+
+    return false;
+}
+
 documentHelper.isViewable = function (addon, extension) {
     const formats = addon.config.docServer().formats || [];
 
@@ -63,20 +73,28 @@ documentHelper.getDocumentKey = function (attachmentId, updateTime) {
 }
 
 documentHelper.getEditorConfig = async function (addon, httpClient, clientKey, localBaseUrl, hostBaseUrl, attachmentInfo, userInfo) {
-
     const fileType = documentHelper.getFileExtension(attachmentInfo.title);
-    let mode = "view";
+
     let callbackUrl = null;
 
     let permittedOperations = await requestHelper.getPermittedOperationsForContent(httpClient, userInfo.accountId, "pages", attachmentInfo.pageId);
+
+    let isEditable = documentHelper.isEditable(addon, fileType);
+    let isFillable = documentHelper.isFillable(addon, fileType);
 
     const permissionEdit = permittedOperations.operations
         .filter(operation => operation.targetType === "attachment" && operation.operation === "create")
         .length > 0;
 
-    if (permissionEdit && documentHelper.isEditable(addon, fileType)) {
-        mode = "edit";
-        callbackUrl = await urlHelper.getCallbackUrl(addon, localBaseUrl, clientKey, userInfo.accountId, attachmentInfo.pageId || attachmentInfo.blogPostId, attachmentInfo.id);
+    if (permissionEdit && (isEditable || isFillable)) {
+        callbackUrl = await urlHelper.getCallbackUrl(
+            addon,
+            localBaseUrl,
+            clientKey,
+            userInfo.accountId,
+            attachmentInfo.pageId || attachmentInfo.blogPostId,
+            attachmentInfo.id
+        );
     }
 
     return {
@@ -93,7 +111,8 @@ documentHelper.getEditorConfig = async function (addon, httpClient, clientKey, l
                 uploaded: attachmentInfo.version.createdAt
             },
             permissions: {
-                edit: permissionEdit,
+                edit: permissionEdit && isEditable,
+                fillForms: permissionEdit && isFillable
             },
             referenceData: {
                 fileKey: attachmentInfo.id,
@@ -102,7 +121,7 @@ documentHelper.getEditorConfig = async function (addon, httpClient, clientKey, l
         },
         editorConfig: {
             callbackUrl: callbackUrl,
-            mode: mode,
+            mode: "edit",
             lang: "en",
             user: {
                 id: userInfo.accountId,